refactor(app): переименовал каталоги(теперь они однотипно пронаименованы) и отформатировал код в компонентах
