+++ package.json
@@ -13,6 +13,7 @@
     "jquery": "^3.5.1",
     "owl.carousel": "^2.3.4",
     "popper.js": "^1.16.1",
+    "prettier": "^2.1.2",
     "react": "^16.13.1",
     "react-dom": "^16.13.1",
     "react-owl-carousel": "^2.3.1",
