--- src/Pages/Blog/Blog.js
@@ -1,19 +1,0 @@
-import React from 'react';
-import './Blog.css';
-
-
-
-
-  export default class Blog extends React.Component {
-    render() {
-        return ( 
-            <div>
-            <h1>Блог</h1>
-        
-            </div>
-        );
-    }
-    }
-
-
-
