--- src/Pages/Home/Home.js
@@ -1,22 +1,0 @@
-import React from 'react';
-import Footer from '../../components/footer/Footer';
-import Banner from '../../components/header/banners/banner';
-
-
-
-
-
-
-
-export default class Home extends React.Component {
-  render () {
-
-    return (
-      <div>
-      <Banner />
-      <Footer /> 
-      </div>
-     
-	)
-  }
-}
