+++ src/components/Footer/Footer.js
@@ -1,0 +1,28 @@
+import React from "react";
+
+export default class Footer extends React.Component {
+  render() {
+    return (
+      <footer>
+        <div class="container">
+          <div class="row">
+            <div class="col-md-12">
+              <div class="inner-content">
+                <p>
+                  Copyright &copy; 2020 Sixteen Clothing Co., Ltd. - Design:{" "}
+                  <a
+                    rel="nofollow noopener"
+                    href="https://templatemo.com"
+                    target="_blank"
+                  >
+                    TemplateMo
+                  </a>
+                </p>
+              </div>
+            </div>
+          </div>
+        </div>
+      </footer>
+    );
+  }
+}
