+++ src/components/Game/Game.js
@@ -1,23 +1,19 @@
-import React from 'react';
-import Board from './Board';
-import './Game.css';
+import React from "react";
+import Board from "./Board";
+import "./Game.css";
 
-  export default class Game extends React.Component {
-    render() {
-      return (
-        <div className="game">
-          <div className="game-board">
-            <Board />
-          </div>
-          <div className="game-info">
-            <div>{/* status */}</div>
-            <ol>{/* TODO */}</ol>
-          </div>
+export default class Game extends React.Component {
+  render() {
+    return (
+      <div className="game">
+        <div className="game-board">
+          <Board />
         </div>
-      );
-    }
+        <div className="game-info">
+          <div>{/* status */}</div>
+          <ol>{/* TODO */}</ol>
+        </div>
+      </div>
+    );
   }
-  
-
- 
-  
+}
