+++ src/components/Game/Square.js
@@ -1,13 +1,10 @@
-import React from 'react';
-import './Game.css';
+import React from "react";
+import "./Game.css";
 
 export default function Square(props) {
-      return (
-        <button
-          className="square"
-          onClick = {props.onClick}>
-          {props.value}
-        </button>
-      );
-    }
-  
+  return (
+    <button className="square" onClick={props.onClick}>
+      {props.value}
+    </button>
+  );
+}
