+++ src/components/Header/Header.js
@@ -1,0 +1,12 @@
+import React from "react";
+import Nav from "../Nav/Nav";
+
+export default class Header extends React.Component {
+  render() {
+    return (
+      <header className="">
+        <Nav />
+      </header>
+    );
+  }
+}
