+++ src/components/Main/Main.js
@@ -1,0 +1,7 @@
+import React from "react";
+
+export default class Main extends React.Component {
+  render() {
+    return <div></div>;
+  }
+}
