--- src/components/header/Header.js
@@ -1,20 +1,0 @@
-import React from 'react';
-import Nav from '../nav/Nav';
-
-
-
-
-
-
-export default class Header extends React.Component {
-  render () {
-    return (
-      <header className="">
-      <Nav />
-     
-      </header>
-      
-    );
-  }
-}
-
