--- src/components/main/Main.js
@@ -1,17 +1,0 @@
-import React from 'react';
-import About from '../../pages/About/About';
-
-
-
-
-  export default class Main extends React.Component {
-     
-    render() {
-
-        return ( 
-          <div></div>
-        )
-    }
-}
-
-
