--- src/components/nav/Nav.js
@@ -1,35 +1,0 @@
-import React from 'react';
-import {Link} from "react-router-dom";
-
-
-
-export default class Nav extends React.Component {
-  render () {
-    return (
-        <nav className="navbar navbar-expand-lg">
-          <div className="container">
-          <Link to ="/" className="navbar-brand"><h2>Sixteen <em>Clothing</em></h2></Link>
-          <button className="navbar-toggler" type="button" data-toggle="collapse" data-target="#navbarResponsive" aria-controls="navbarResponsive" aria-expanded="false" aria-label="Toggle navigation">
-            <span className="navbar-toggler-icon"></span>
-          </button>
-          <div className="collapse navbar-collapse" id="navbarResponsive">
-            <ul className="navbar-nav ml-auto">
-              <li className="nav-item">
-                <Link to="/" className ="nav-link">Home <span class="sr-only">(current)</span></Link>
-              </li>
-              <li className="nav-item">
-                <Link to="/our-products" className ="nav-link">Our Products</Link>
-              </li>
-               <li className="nav-item">
-                <Link to="/about" className ="nav-link">About Us</Link>
-              </li>
-              <li className="nav-item">
-                <Link to="/contact" className ="nav-link">Contact Us</Link>
-              </li>
-            </ul>
-            </div>
-            </div>
-          </nav>
-    );
-  }
-}
