--- src/components/our_team/specialistsData.js
@@ -1,58 +1,0 @@
-import team01 from '../../assets/images/team_01.jpg';
-import team02 from '../../assets/images/team_02.jpg';
-import team03 from '../../assets/images/team_03.jpg';
-import team04 from '../../assets/images/team_04.jpg';
-import team05 from '../../assets/images/team_05.jpg';
-import team06 from '../../assets/images/team_06.jpg';
-
-const specialistsData = [
-    {
-        id: 1,
-        h4: "Johnny William",
-        specialization:"CO-Founder",
-        description: "Lorem ipsum dolor sit amet, consectetur adipisicing itaque corporis nulla.",
-        img:team01
-    },
-
-    {
-        id: 2,
-        h4: "Karry Pitcher",
-        specialization: "Product Expert",
-        description:"Lorem ipsum dolor sit amet, consectetur adipisicing itaque corporis nulla.",
-        img: team02
-    },
-
-    {
-        id: 3,
-        h4: "Michael Soft",
-        specialization: "Chief Marketing",
-        description: "Lorem ipsum dolor sit amet, consectetur adipisicing itaque corporis nulla.",
-        img: team03
-    },
-
-    {
-        id: 4,
-        h4: "Mary Cool",
-        specialization: "Product Specialist",
-        description: "Lorem ipsum dolor sit amet, consectetur adipisicing itaque corporis nulla.",
-        img:team04
-    },
-
-    {
-        id: 5,
-        h4: "George Walker",
-        specialization: "Product Photographer",
-        description: "Lorem ipsum dolor sit amet, consectetur adipisicing itaque corporis nulla.",
-        img: team05
-    },
-
-    {
-        id: 6,
-        h4: "Kate Town",
-        specialization: "General Manager",
-        description:"Lorem ipsum dolor sit amet, consectetur adipisicing itaque corporis nulla.",
-        img: team06
-    }
-]
-
-export default specialistsData;
