+++ src/pages/Blog/Blog.js
@@ -1,0 +1,12 @@
+import React from "react";
+import "./Blog.css";
+
+export default class Blog extends React.Component {
+  render() {
+    return (
+      <div>
+        <h1>Блог</h1>
+      </div>
+    );
+  }
+}
